Add logout
